Move Climacell into weather providers
